lunch_decision/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django
flake8